+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/LibrarySelectionTypeDto.java
@@ -1,32 +1,32 @@
 package uk.ac.bbsrc.tgac.miso.dto;
 
 public class LibrarySelectionTypeDto {
-  private String alias;
-  private long id;
+  private Long id;
   private String name;
+  private String description;
 
-  public String getAlias() {
-    return alias;
+  public Long getId() {
+    return id;
   }
 
-  public long getId() {
-    return id;
+  public void setId(Long id) {
+    this.id = id;
   }
 
   public String getName() {
     return name;
   }
 
-  public void setAlias(String alias) {
-    this.alias = alias;
-  }
-
-  public void setId(long id) {
-    this.id = id;
-  }
-
   public void setName(String name) {
     this.name = name;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
   }
 
 }