+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/MenuController.java
@@ -102,6 +102,7 @@ import uk.ac.bbsrc.tgac.miso.service.KitDescriptorService;
 import uk.ac.bbsrc.tgac.miso.service.LabService;
 import uk.ac.bbsrc.tgac.miso.service.LibraryDesignCodeService;
 import uk.ac.bbsrc.tgac.miso.service.LibraryDesignService;
+import uk.ac.bbsrc.tgac.miso.service.LibrarySelectionService;
 import uk.ac.bbsrc.tgac.miso.service.LibraryService;
 import uk.ac.bbsrc.tgac.miso.service.PartitionQcTypeService;
 import uk.ac.bbsrc.tgac.miso.service.QualityControlService;
@@ -156,6 +157,8 @@ public class MenuController implements ServletContextAware {
   private TissueTypeService tissueTypeService;
   @Autowired
   private LibraryService libraryService;
+  @Autowired
+  private LibrarySelectionService librarySelectionService;
   @Autowired
   private LibraryDesignService libraryDesignService;
   @Autowired
@@ -288,7 +291,7 @@ public class MenuController implements ServletContextAware {
 
     createArray(mapper, node, "libraryDesigns", libraryDesignService.list(), Dtos::asDto);
     createArray(mapper, node, "libraryTypes", libraryService.listLibraryTypes(), Dtos::asDto);
-    createArray(mapper, node, "librarySelections", libraryService.listLibrarySelectionTypes(), Dtos::asDto);
+    createArray(mapper, node, "librarySelections", librarySelectionService.list(), Dtos::asDto);
     createArray(mapper, node, "libraryStrategies", libraryService.listLibraryStrategyTypes(), Dtos::asDto);
     createArray(mapper, node, "libraryDesignCodes", libraryDesignCodeService.list(), Dtos::asDto);
     Set<Long> activePlatforms = sequencerService.list().stream().filter(Instrument::isActive)