+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 Changes:
 
   * Added interface for administrators to upload and delete location maps
+  * Added interface for administrators to add, edit, and delete dropdown options
+    * library selection types
   * Changed the way location maps are specified for freezers
 
 Upgrade Notes: